+++ src/main/java/error/Exception400.java
@@ -1,7 +1,13 @@
 package error;
 
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.BAD_REQUEST)
 public class Exception400 extends RuntimeException {
     public Exception400(String msg) {
         super(msg);
     }
+
+
 }
