+++ src/main/java/roomescape/reservation/Reservation.java
@@ -1,0 +1,42 @@
+package roomescape.reservation;
+
+
+import java.util.concurrent.atomic.AtomicLong;
+
+public class Reservation {
+
+    private Long id;
+    private String name;
+    private String date;
+    private String time;
+
+    private static AtomicLong index = new AtomicLong(1 );
+
+    public static Reservation toEntity(Reservation reservation) {
+        Reservation reservation1 = new Reservation(index.getAndIncrement(), reservation.getName(), reservation.getDate(), reservation.getTime());
+        return reservation1;
+    }
+
+
+    public String getName() {
+        return name;
+    }
+    public long getId() {
+        return id;
+    }
+
+    public String getDate() {
+        return date;
+    }
+
+    public String getTime() {
+        return time;
+    }
+
+    public Reservation(Long id, String name, String date, String time) {
+        this.id = id;
+        this.name = name;
+        this.date = date;
+        this.time = time;
+    }
+}
